Add files via upload
Data and R script used for project. R script includes all packages and code used for the project. Original dataset was sourced from ChIP Atlas and was ommitted due to file size (>2Gb). Original dataset contained all TF data for the hg38 human genome in cardiovascular cell lines. This was cleaned and processed to produce the TFsPXDN and TFsPXDNL files.